ingestion/requirements.txt
solana-py
backoff
websockets
psycopg2-binary
redis
python-dotenv
asyncio
aiohttp
